Move validateEmail to helpers/Contact for later use
