+++ src/renderer/components/helpers/Contact.js
@@ -7,6 +7,26 @@ export function renderAvatar (avatarPath, color, displayName) {
   )
 }
 
+export function isValidEmail (email) {
+  // empty string is not allowed
+  if (email === '') return false
+  const parts = email.split('@')
+  // missing @ character or more than one @ character
+  if (parts.length !== 2) return false
+  const [local, domain] = parts
+  // empty string is not valid for local part
+  if (local === '') return false
+  // domain is too short
+  if (domain.length <= 3) return false
+  const dot = domain.indexOf('.')
+  // invalid domain without a dot
+  if (dot === -1) return false
+  // invalid domain if dot is (second) last character
+  if (dot >= domain.length - 2) return false
+
+  return true
+}
+
 export function Avatar (props) {
   const { avatarPath, color, displayName } = props
   if (avatarPath) {
